requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
ca-certificates=2023.5.7=hbcca054_0
certifi=2023.5.7=pypi_0
charset-normalizer=3.2.0=pypi_0
idna=3.4=pypi_0
imageio=2.31.1=pypi_0
joblib=1.3.1=pypi_0
ld_impl_linux-64=2.40=h41732ed_0
libffi=3.4.2=h7f98852_5
libgcc-ng=13.1.0=he5830b7_0
libgomp=13.1.0=he5830b7_0
libnsl=2.0.0=h7f98852_0
libsqlite=3.42.0=h2797004_0
libstdcxx-ng=13.1.0=hfd8a6a1_0
libzlib=1.2.13=hd590300_5
ncurses=6.4=hcb278e6_0
networkx=2.6.3=pypi_0
numpy=1.21.6=pypi_0
opencv-python=4.8.0.74=pypi_0
openssl=3.1.1=hd590300_1
packaging=23.1=pypi_0
pillow=9.5.0=pypi_0
pip=23.2=pyhd8ed1ab_0
python=3.7.12=hf930737_100_cpython
pywavelets=1.3.0=pypi_0
readline=8.2=h8228510_1
requests=2.31.0=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
setuptools=68.0.0=pyhd8ed1ab_0
sqlite=3.42.0=h2c6b66d_0
threadpoolctl=3.1.0=pypi_0
tifffile=2021.11.2=pypi_0
tk=8.6.12=h27826a3_0
torch=1.12.1+cu113=pypi_0
torchaudio=0.12.1+cu113=pypi_0
torchvision=0.13.1+cu113=pypi_0
tqdm=4.65.0=pypi_0
typing-extensions=4.7.1=pypi_0
urllib3=2.0.3=pypi_0
wheel=0.40.0=pyhd8ed1ab_1
xz=5.2.6=h166bdaf_0
